+++ packages/common/src/cache/memory.ts
@@ -1,15 +1,10 @@
 import { isRegExp } from 'util/types';
 
-import { array, demethodize, maybe, Option } from '@fluss/core';
+import { demethodize, list, maybe, Option } from '@fluss/core';
 
-export interface MemoryCache<Key, Value> {
-  readonly put: (key: Key, value: Value) => void;
-  readonly has: (key: Key | RegExp) => boolean;
-  readonly get: (key: Key | RegExp) => Option<Value>;
-  readonly clear: () => void;
-  readonly remove: (key: Key | RegExp) => void;
-  readonly through: (key: Key, or: () => Value) => Value;
-}
+import { Cache } from './cache';
+
+export interface MemoryCache<Key, Value> extends Cache<Key, Value> {}
 
 export const initMemoryCache = <Key extends string, Value>(): MemoryCache<
   Key,
@@ -17,22 +12,18 @@ export const initMemoryCache = <Key extends string, Value>(): MemoryCache<
 > => {
   const cache = new Map<Key, Value>();
 
-  const findKey = (value: Key | RegExp): Option<Key> => {
-    const keys = array(cache.keys());
-
-    return maybe(
-      keys.find((key) =>
-        isRegExp(value)
-          ? value.test(key)
-          : value.includes(key) || key.includes(value),
-      )!,
+  const findKey = (value: Key | RegExp): Option<Key> =>
+    list(cache.keys()).find((key) =>
+      isRegExp(value)
+        ? value.test(key)
+        : value.includes(key) || key.includes(value),
     );
-  };
 
   return {
     get: (key) => findKey(key).chain((key) => maybe(cache.get(key))),
     has: (key) => findKey(key).isSome(),
     put: demethodize(cache, 'set'),
+    keys: () => list(cache.keys()),
     clear: demethodize(cache, 'clear'),
     remove: (key) => findKey(key).map((key) => cache.delete(key)),
     through: (key, or) =>
@@ -44,5 +35,7 @@ export const initMemoryCache = <Key extends string, Value>(): MemoryCache<
           return value;
         })
         .extract(),
+    entries: () => list(cache.entries()),
+    isEmpty: () => cache.size === 0,
   };
 };